[Curves] Disable slider/button for 2D render and add typedoc (#49)
* Disable slider and play button on 2D renders

* Add function.ts for curves

* Fix eslint for naming convention

* Add examples to basic curve functions

* Improve code quality

* Add curves to README.md
